refactor: add typescript support
