+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["env", "react"]
+  "presets": ["@babel/env", "@babel/react", "@babel/typescript"]
 }
