+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Header from './Header.jsx';
+import Header from './Header';
 
 const App = () => {
   return (
