+++ src/index.ts
@@ -2,5 +2,6 @@ import './css/body.css';
 import './css/app.css';
 import './css/header.css';
 
-import App from './App.jsx';
+
+import './App';
